Added add'l logic to setting up symlinks
Now we check to see if the dotfile exists, if it does, then we check if
it is already set as the proper symlink to our dotfiles. If it is not
properly set, we then backup (mv) the existing file or directory to
.dotfiles_old, then we create the dotfile symlink as it should be.

Using this new method keeps from filling up our .dotfiles_old directory
with a bunch of backed up dotfiles which are nothing more than symlinks
to our desired .dotfiles location. The previous method was not efficient
and purely just did a backup (mv) no matter what and then recreated the
symlink.